+++ sc-tools/sc-builder/src/builder.hpp
@@ -21,6 +21,10 @@ struct BuilderParams
   std::string m_enabledExtPath;
   //! Flag to generate format information based on file extensions
   bool m_autoFormatInfo : 1;
+  //! Result structure system identifier
+  std::string m_resultStructureSystemIdtf;
+  //! Flag to create result structure
+  bool m_resultStructureUpload = SC_FALSE;
 };
 
 class Builder
@@ -42,5 +46,6 @@ private:
   std::list<std::string> m_files;
 
   BuilderParams m_params;
+  ScAddr m_outputStructure;
   std::unique_ptr<ScMemoryContext> m_ctx;
 };