+++ sc-tools/sc-builder/tests/units/test_base.cpp
@@ -6,6 +6,7 @@
 
 #include "sc_options.hpp"
 #include "sc_memory_config.hpp"
+#include "sc_builder_config.hpp"
 
 TEST(ScBuilder, RunStop)
 {
@@ -22,10 +23,78 @@ TEST(ScBuilder, RunStop)
   memoryParams.insert({"repo_path", SC_BUILDER_REPO_PATH});
   memoryParams.insert({"clear", "true"});
 
-  ScMemoryConfig memoryConfig{config, {"repo_path"}, std::move(memoryParams)};
+  ScConfig configFile{config, {"repo_path"}};
+  std::string memoryGroupName = "sc-memory";
+
+  ScMemoryConfig memoryConfig{configFile, std::move(memoryParams)};
 
   Builder builder;
   EXPECT_TRUE(builder.Run(params, memoryConfig.GetParams()));
+
+  ScParams scParams{options, {}};
+  if (configFile.IsValid())
+  {
+    ScConfigGroup group = configFile[memoryGroupName];
+    for (auto const & key : *group)
+    {
+      std::string const & value = group[key];
+      std::stringstream stream;
+
+      scParams.insert({key, value});
+    }
+    sc_memory_params const scMemoryParams = memoryConfig.GetParams();
+
+    EXPECT_EQ(std::stoi(scParams.at("max_loaded_segments")), scMemoryParams.max_loaded_segments);
+    EXPECT_EQ(std::stoi(scParams.at("max_threads")), scMemoryParams.max_threads);
+    EXPECT_EQ(std::stoi(scParams.at("save_period")), scMemoryParams.save_period);
+    EXPECT_EQ(std::stoi(scParams.at("update_period")), scMemoryParams.update_period);
+    EXPECT_EQ(scParams.at("log_type"), scMemoryParams.log_type);
+    EXPECT_EQ(scParams.at("log_file"), scMemoryParams.log_file);
+    EXPECT_EQ(scParams.at("log_level"), scMemoryParams.log_level);
+  }
+}
+
+TEST(ScBuilder, BuilderConfig)
+{
+  ScOptions options{1, nullptr};
+
+  BuilderParams params;
+  params.m_inputPath = SC_BUILDER_KB;
+  params.m_outputPath = SC_BUILDER_REPO_PATH;
+  params.m_autoFormatInfo = SC_TRUE;
+
+  std::string config = SC_BUILDER_INI;
+
+  ScParams memoryParams{options, {}};
+  memoryParams.insert({"repo_path", SC_BUILDER_REPO_PATH});
+  memoryParams.insert({"clear", "false"});
+
+  ScConfig configFile{config, {"repo_path"}};
+
+  ScBuilderConfig builderConfig{configFile, std::move(params)};
+  ScMemoryConfig memoryConfig{configFile, std::move(memoryParams)};
+
+  Builder builder;
+  BuilderParams builderParams = builderConfig.GetParams();
+  EXPECT_TRUE(builder.Run(builderParams, memoryConfig.GetParams()));
+
+  std::string builderGroupName = "sc-builder";
+
+  ScParams scParams{options, {}};
+  if (configFile.IsValid())
+  {
+    ScConfigGroup group = configFile[builderGroupName];
+    for (auto const & key : *group)
+    {
+      std::string const & value = group[key];
+      std::stringstream stream;
+
+      scParams.insert({key, value});
+    }
+
+    EXPECT_EQ(scParams.at("result_structure_upload") == "true" ? true : false, builderParams.m_resultStructureUpload);
+    EXPECT_EQ(scParams.at("result_structure_system_idtf"), builderParams.m_resultStructureSystemIdtf);
+  }
 }
 
 TEST_F(ScBuilderTest, Smoke)
@@ -33,11 +102,11 @@ TEST_F(ScBuilderTest, Smoke)
   ScMemoryContext ctx("Builder_Base");
 
   std::string const scsData =
-    "base"
-    " _<- test_element;"
-    " _=> nrel_main_idtf::"
-    "    _[]"
-    "    (* _<- lang_en;; *);;";
+      "base"
+      " _<- test_element;"
+      " _=> nrel_main_idtf::"
+      "    _[]"
+      "    (* _<- lang_en;; *);;";
 
   ScTemplate templ;
   EXPECT_TRUE(ctx.HelperBuildTemplate(templ, scsData));